requirements.txt
# Core dependencies
numpy==1.24.3
pandas==2.1.0
matplotlib==3.7.2
plotly==5.15.0
scikit-learn==1.3.0
python-dotenv==1.0.0
streamlit==1.27.0

# API and data fetching
ccxt==4.1.13
requests==2.31.0
websockets==11.0.3
tweepy==4.14.0
praw==7.7.1
newsapi-python==0.2.7

# Machine learning
torch==2.0.1
transformers==4.33.2
nltk==3.8.1
textblob==0.17.1

# Technical analysis
ta==0.10.2
pandas-ta==0.3.14b

# Database
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
redis==5.0.1

# Dashboard
streamlit==1.25.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.66.1
joblib==1.3.2
